+++ src/ulib/orm/driver/orm_driver_sqlite.cpp
@@ -82,8 +82,8 @@ void UOrmDriverSqlite::handlerError()
    if (UOrmDriver::errmsg  == 0) UOrmDriver::errmsg  = U_SYSCALL(sqlite3_errmsg,  "%p", (sqlite3*)UOrmDriver::connection);
    if (UOrmDriver::errcode == 0) UOrmDriver::errcode = U_SYSCALL(sqlite3_errcode, "%p", (sqlite3*)UOrmDriver::connection);
 
-   if (UOrmDriver::errcode >= 0                                &&
-       UOrmDriver::errcode < U_NUM_ELEMENTS(error_value_table) &&
+   if (UOrmDriver::errcode >= 0                                     &&
+       UOrmDriver::errcode < (int)U_NUM_ELEMENTS(error_value_table) &&
        UOrmDriver::errcode == error_value_table[UOrmDriver::errcode].value)
       {
       UOrmDriver::errname = error_value_table[UOrmDriver::errcode].name;
@@ -92,7 +92,7 @@ void UOrmDriverSqlite::handlerError()
       {
       UOrmDriver::errname = "???";
 
-      for (int i = 0; i < U_NUM_ELEMENTS(error_value_table); ++i)
+      for (unsigned int i = 0; i < U_NUM_ELEMENTS(error_value_table); ++i)
          {
          if (UOrmDriver::errcode == error_value_table[i].value)
             {
